+++ beacon-chain/p2p/discovery.go
@@ -349,7 +349,7 @@ func (s *Service) listenForNewNodes() {
 				wg.Add(1)
 				go func(info *peer.AddrInfo) {
 					if err := s.connectWithPeer(s.ctx, *info); err != nil {
-						log.WithError(err).Tracef("Could not connect with peer %s", info.String())
+						log.WithError(err).WithField("peerID", info.ID).Debug("Could not connect with peer")
 					}
 					wg.Done()
 				}(peerInfo)