+++ beacon-chain/sync/initial-sync/blocks_queue.go
@@ -337,8 +337,10 @@ func (q *blocksQueue) onDataReceivedEvent(ctx context.Context) eventHandlerFn {
 			}
 			if errors.Is(response.err, beaconsync.ErrInvalidFetchedData) {
 				// Peer returned invalid data, penalize.
-				q.blocksFetcher.p2p.Peers().Scorers().BadResponsesScorer().Increment(m.pid)
-				log.WithField("pid", response.pid).Debug("Peer is penalized for invalid blocks")
+				score := q.blocksFetcher.p2p.Peers().Scorers().BadResponsesScorer().Increment(m.pid)
+				log.
+					WithFields(logrus.Fields{"pid": response.pid, "newBadResponsesScore": score}).
+					Debug("Peer is penalized for invalid blocks")
 			}
 			return m.state, response.err
 		}